ready interface and implement for BeaconMenager

+++ app/src/main/java/com/mrx/indoorservice/domain/model/BeaconsEnvironmentInfo.kt
@@ -1,3 +1,3 @@
 package com.mrx.indoorservice.domain.model
 
-data class BeaconsEnvironmentInfo(var beaconId: String, var distance: Float)
+data class BeaconsEnvironmentInfo(var beaconId: String, var distance: Double)
